Refatorando as paginas index e listar
