+++ Pages/Produtos/Create.cshtml
@@ -48,12 +48,12 @@
         <div class="botao-cadastro">
             <div class="entrada-cadastro">
                 <div class="form-group">
-                    <button type="button" class="btn btn-primary btn-sm w-100 form-cadastro-botao">Cadastrar</button>
+                    <button type="submit" class="btn btn-primary btn-sm w-100 form-cadastro-botao">Cadastrar</button>
                 </div>
             </div>
             <div class="entrada-cadastro">
                 <div class="form-group">
-                    <button type="button" class="btn btn-outline-secondary btn-sm w-100 form-cadastro-botao-cancelar">Cancelar</button>
+                    <button type="button" class="btn btn-outline-secondary btn-sm w-100 form-cadastro-botao-cancelar" asp-page="/Produtos/Index">Cancelar</button>
                 </div>
             </div>
         </div>
