+++ Pages/Produtos/Index.cshtml
@@ -1,28 +1,27 @@
 ﻿@page
-@model InDEVstockWebApp.Pages.Produtos.IndexModel;
+@model InDEVstockWebApp.Pages.Produtos.PaginaInicialModel
 @{
 }
-<div class="container">
-    <h1 class="text-center">Lista de Produtos</h1>
-    <section>
-        @foreach (var data in Model.Produtos)
-        {<div class="produto-picture">
-                <img src="@Html.DisplayFor(modelItem => data.image)"
-                     alt="Imagem do @Html.DisplayFor(modelItem => data.Nome)" />
-            </div>
-         <div class="Produtos-info">
-             <h1>Nome: @Html.DisplayFor(modelItem => data.Nome)</h1>
-             <h2>Id: @Html.DisplayFor(modelItem => data.Id)</h2>
-             <h2>Modelo: @Html.DisplayFor(modelItem => data.Modelo)</h2>
-             <h3>Cor: @Html.DisplayFor(modelItem => data.Cor)</h3>
-             <h3>Tamanho: @Html.DisplayFor(modelItem => data.Tamanho)</h3>
-             <h3>Quantidade: @Html.DisplayFor(modelItem => data.Quantidade)</h3>
-         </div>
-            <a class="btn btn-info"
-               asp-page="/Produtos/Detail"
-               asp-route-id="@data.Id">Detalhes</a>
-        }
 
-        }
-    </section>
-</div>
+<html>
+    <head>
+        <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
+        <link rel="stylesheet" href="~/css/site.css" />
+        <link rel="stylesheet" href="~/css/StyleSheet.css" />
+    </head>
+
+    <body>
+        <div class="container">
+            <div class="row">
+                <div class="col">
+                    <a asp-page="/Produtos/Create" class="btn btn-primary form-control btn-block">Cadastrar produto</a>
+                </div>
+                <div class="col">
+
+                    <a asp-page="/Produtos/ListaProdutos" class="btn btn-primary form-control btn-block">Listar Produto</a>
+                </div>
+            </div>
+        </div>
+    </body>
+
+</html>
