+++ Pages/Produtos/Index.cshtml.cs
@@ -4,35 +4,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using InDEVstockWebApp.Model;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using Newtonsoft.Json;
-
 
 namespace InDEVstockWebApp.Pages.Produtos
 {
-    public class IndexModel : PageModel
+    public class PaginaInicialModel : PageModel
     {
-        public List<Produto> Produtos { get; private set; }
-        string baseUrl = "http://localhost:49268"; 
-        public async Task OnGetAsync()
+        public void OnGet()
         {
-            using (var client = new HttpClient())
-            {
-                client.BaseAddress = new Uri(baseUrl);
-                client.DefaultRequestHeaders.Clear();
-                client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-                HttpResponseMessage response = await client.GetAsync("api/Produtos");
-                if (response.IsSuccessStatusCode)
-                {
-                    var result = response.Content.ReadAsStringAsync().Result;
-                    Produtos = JsonConvert.DeserializeObject<List<Produto>>(result);
-
-                }
-            }
-
         }
     }
 }
-
