+++ Pages/Shared/_Layout.cshtml
@@ -12,35 +12,15 @@
 <body>
 
     <header>
-
-
         <div class=" text-center mt-2">
-
             <h1 class="pt-3 pb-3 bg-dark text-white">InDevStock </h1>
             <h2 class="pt-3 pb-3">A forma mais inteligente de gerenciar o seu estoque de roupas!!</h2>
             <hr />
-
         </div>
-
-
     </header>
 
-    <div class="container">
-        <div class="row">
-            <div class="col">
-                <a asp-page="/Produtos/Index" class="btn btn-primary form-control btn-block">Cadastrar produto</a>
-            </div>
-            <div class="col">
-
-                <a asp-page="/Produtos/Index" class="btn btn-primary form-control btn-block">Listar Produto</a>
-            </div>
-        </div>
-    </div>
-
-    <div class="text-center mt-5 mb-5">
-
+    <div class=" text-center mt-5 mb-5">
         <img src ="https://gcdn.pbrd.co/images/Tuk1eg3JJDeN.webp?o=1"/>
- 
     </div>
